make GatAllAppsRequestHandler work. But with question...
